renamed components and fixed storybook bugs
